adds common ConnectionManager to reduce code duplication
